Aqualung i770R: _really_ fix the gas mix parsing
Commit fac3c8ff1487 ("Aqualung i770R: fix up gas mix parsing") was
supposed to fix gasmix parsing for the i770R, but because Janice's patch
didn't apply as-is, I had just redone it by hand, and in the process
copied the wrong o2_offset for the i770R, giving (not surprisingly) the
wrong results.

This just fixes my copy-paste error to what Janice actually had
originally.

Mea culpa.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>