Suunto EON Steel: use the generic field cache infrastructure
The only real changes here are some interface updates: several traversal
routines are changed to return 'dc_status_t' which makes more sense in
the libdivecomputer world anyway, and that matches the generic field
cache string routines.

The other changes are also a direct result of the generic code using
slightly different names for the cache fields: they use the same names
as the DC_FIELD_xyz enum to fetch them.

This is not a full conversion to the generic wrappers, and the EON Steel
parser accesses the cache structure directly in several places because
of how the code was written.  That's fine.  Not pretty, but this is all
totally internal to libdivecomputer.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>