iostream: protect write side against bad custom IO interfaces too
This is the same as the previous commit, just for the write side error
handling.

If we have a bad custom IO low-level driver that returns
DC_STATUS_SUCCESS with zero bytes written, and we have a write() user
that isn't ready for partial writes, we just consider that an IO error.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>