OSTC parser whitespace cleanups
This is pure whitespace cleanup, where the hw_ostc_parser.c file had a
number of lines with eight spaces instead of tabs.  They were right next
to lines with tabs, so it really made no sense.

I only noticed because some of it was arbitrary difference to Jef's
upstream code, and it stood out like a sore thumb when trying to
re-generate our local changes wrt upstream.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>