shearwater: remove unused 'shearwater_common_command()' function
Dirk added this helper in commit 9379004b2df5 ("Shearwater: add helper
to send bytes to the dive computer"), and it was used for a short while
by commit 437cc3e0cc9f ("Shearwater: try to gracefully shut down the
Bluetooth connection").

But then the use was reverted in commit 8120b11258d8, and this helper
function is now stale and unused, and an unnecessary difference to Jef's
upstream code base.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>