Mark the Aqualung i750TC as BLE-capable
Dirk got a log from Amin that shows

  Discovered new device: "EZ001945" "LE:{cb0ff33a-40d3-63eb-2764-11e63896534f}"
  Not recognized as dive computer

and that's the i750TC device model name.  It's not recognized as a dive
computer because subsurface don't have that EZ as a model version, but
even if it did, it would not try to use BLE to download because
libdivecomputer doesn't mark it as BLE capable.

But it appears that it is, so add the protocol marker (even if the log
Dirk has then shows a failure to connect, so not all is well).

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>