EON Steel: support new gas change data from newer firmware
It seems that Suunto extended the gas change information in FW 2.5, and
added events and fields for when you create and delete gas mixes under
water.

The parses the relevant new information to get gas switches working
again, and to also add the information about gas creation and deletion
events.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>