Remove MSVC build instructions from github workflows
MSVC isn't a supported target for subsurface - the Windows build is
cross-compiled, not native.  And MSVC is actively user-hostile, with
insane errors and warnings.

In particular, MSVC doesn't like "strdup()", and suggests you use the
nonstandard _strdup() instead.  Which is all kinds of wrong, and seems
to be a "let's use strict POSIX namespace rules as an excuse to make
people write less portable code".

There is probably some flag to make the MSVC compiler happy with sane
code, but it's easier to just disable the insanity entirely.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>