Add Shearwater Perdix 2 and Petrel 3
Also update and fix the hardware IDs based on the latest information
from Shearwater.
Two of the hardware IDs that we supported before are not listed in the
current docs. In an abundance of caution I'll leave them in the code.
No harm done if they don't exist in real life.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>