Added detection for bailout to OC / switch back to CC for CCR dives.
As a CCR diver I would like to see the ability to properly track the 'on
loop' / 'bailed out to open circuit' status directly in Subsurface,
because this is an important (or even the most important) bit of status
information during a CCR dive.  This should extend to the use of the
correct ppO2 / gas mix in the deco ceiling / tissue model calculation.

My idea for how to do this would be to track the 'type' ('diluent' / 'OC
bailout') for every gasmix / tank that is reported by the dive computer.

Most CCR capable dive computers that I am familiar with require the user
to enter two different gas lists for diluent and bailout, so this should
work with the existing libdivecomputer API for these.  Unfortunately I
think making this change in Subsurface will require a bit of work, as
the libdivecomputer field capable of tracking the 'type' of a gas or
tank (`cache->tankinfo[]`) does not seem to be consumed at all in
Subsurface.

So this just provides a prerequisite for the change in Subsurface by
populating `tankinfo[]`.  In addition to this it also triggers a message
on every switch from CC to OC and back, at least giving a visual
indication of these diver triggered events.  The messages can probably
be removed from libdivecomputer again once 'loop status' tracking has
been added to Subsurface.

Also included is a fix of the tab expansion mess that I created in
commit 2129403 ("Added facility to detect and interpret manual setpoint
switches for Garmin Descent computers").  Apologies for this, I've
switched to using a custom `.vimrc` for this project now.

[ Edited up the commit message a bit further   - Linus ]

Signed-off-by: Michael Keller <github@ike.ch>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>