refactoring: Push field cache access down into base class.
Push accesses to cached fields down into `field-cache.c` from parsers
that use the field cache.

As a result, all cached fields will be available to Subsurface in
computer models that use the field cache.

In particular this means that the proper use type (diluent / OC bailout)
of tanks will be shown for the Garmin Descent dive computers once that
is merged.

[ Edited commit message and massaged the deepblu and Eon Steel parsers
  to not complain about unhandled switch statement cases. - Linus ]

Signed-off-by: Michael Keller <github@ike.ch>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>