garmin parser: avoid build warning about converting pointer types
DECLARE_FIELD() uses array_uint_endian() to turn an integer type into
the right endianness.  It's all conditional on being an integer type,
but the compiler still sees the assignment (with a cast) for other
types, and complains about casting the 'unsigned int' return value to a
pointer, even when that case is not actually dynamically ever taken.

Fix the compiler warning by just changing the return type of this
conversion function to 'unsigned long' instead, which will match the
size of pointers on all relevant architectures.

Don't look at that macro too closely, you'll go blind.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>