garmin_parser: add a HRM profile message stub
This is an empty stub for the HRM profile message, which isn't actually
used by anything I know of, but came up as a result of some mis-parsing
of odd FIT files generated by the Suunto mobile app.

Losing synchronization in the FIT file then caused the parser to think
it needed this message type, and not having it then caused an early
abort.

While it's not actually needed once parsing things correctly, since I
looked up the message number and name for this message type, let's just
keep it around.  It won't hurt, and maybe it avoids me having to look it
up in the future.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>